+++ Modules/Loadable/Annotations/MRML/vtkMRMLAnnotationAngleStorageNode.cxx
@@ -77,7 +77,7 @@ int vtkMRMLAnnotationAngleStorageNode::ReadAnnotationAngleData(vtkMRMLAnnotation
 
   vtkDebugMacro("got a line: \n\"" << line << "\""); 
   std::string attValue(line);
-  int size = std::string(this->GetAnnotationStorageType()).size();
+  size_t size = std::string(this->GetAnnotationStorageType()).size();
  
   if (attValue.compare(0,size,this->GetAnnotationStorageType()))
     {
@@ -152,7 +152,7 @@ int vtkMRMLAnnotationAngleStorageNode::ReadAnnotationAngleProperties(vtkMRMLAnno
   // if there's a space after the hash, try to find options
 
  std::string preposition = std::string("# ") + this->GetAnnotationStorageType();
- int pointOffset = preposition.size();
+ size_t pointOffset = preposition.size();
 
  vtkDebugMacro("Have a possible option in line " << line);
  std::string lineString = std::string(line);
