+++ Modules/Loadable/Annotations/MRML/vtkMRMLAnnotationLinesStorageNode.cxx
@@ -67,7 +67,7 @@ int vtkMRMLAnnotationLinesStorageNode::ReadAnnotationLineDisplayProperties(vtkMR
       return flag; 
     }
 
-  int pointOffset = preposition.size();
+  size_t pointOffset = preposition.size();
   preposition.insert(0,"# ");
 
   if (lineString.find(preposition + "LineThickness = ") != std::string::npos)
@@ -108,7 +108,7 @@ int vtkMRMLAnnotationLinesStorageNode::ReadAnnotationLinesData(vtkMRMLAnnotation
 
   vtkDebugMacro("got a line: \n\"" << line << "\""); 
   std::string attValue(line);
-  int size = std::string(this->GetAnnotationStorageType()).size();
+  size_t size = std::string(this->GetAnnotationStorageType()).size();
  
   if (attValue.compare(0,size,this->GetAnnotationStorageType()))
     {
