+++ Modules/Loadable/Annotations/MRML/vtkMRMLAnnotationRulerStorageNode.cxx
@@ -79,7 +79,7 @@ int vtkMRMLAnnotationRulerStorageNode::ReadAnnotationRulerData(vtkMRMLAnnotation
   vtkDebugMacro("got a line: \n\"" << line << "\""); 
 
   std::string attValue(line);
-  int size = std::string(this->GetAnnotationStorageType()).size();
+  size_t size = std::string(this->GetAnnotationStorageType()).size();
  
   if (attValue.compare(0,size,this->GetAnnotationStorageType()))
     {
@@ -148,7 +148,7 @@ int vtkMRMLAnnotationRulerStorageNode::ReadAnnotationRulerProperties(vtkMRMLAnno
  vtkDebugMacro("Comment line, checking:\n\"" << line << "\"");
 
  std::string preposition = std::string("# ") + this->GetAnnotationStorageType();
- int pointOffset = preposition.size();
+ size_t pointOffset = preposition.size();
 
  //this->DebugOn(); 
  vtkDebugMacro("Have a possible option in line " << line);
